delete useless file.
